requirements.txt
Django>=2.2
djangorestframework>=3.13.1
python-decouple>=3.6
dj-database-url>=0.5.0
drf_spectacular
psycopg2-binary
django-cors-headers==3.13.0
pymysql==1.0.2

